feat(partie): allow player to use near buildings
Subtract the input resources of the building from the player resources.
Add the output resources of the building to the player resources.
Take one coin from the player if he don't own the building used, then give one coin to the player who own it.
Style a bit more precisely the building cells.