feat(partie): allow players to finish a round and start a new one
Ask each player to choose how to feed their workers.
Show only the options that the player can choose (don't propose 'poisson' if the player hasn't any).
Allow player to use piece or score point to pay.
Clean the board from workers when a new round is started.
Set the number of workers to the max allowed for each player.