refacto(joueurs): change the method of getting players data
The app subscribes only once to the players data changes (in JoueursService) and store them in a BehaviorSubject.
All components that need the data subscribe to the BehaviorSubject of the service.
Now that the data are stored in the service, some functions of JoueursService don't need to get them from the caller anymore, so their list of args is shortened.

refacto(joueurs): change args of some functions

Joueurs and JoueurActif are now stored in the service.
Some functions don't have to ask for the list of players nor the active player anymore.